git-svn-id: http://dizzy.googlecode.com/svn/trunk@538 2420a164-f031-0410-b5c9-952b3439b043

+++ app/helpers/comments_helper.rb
@@ -14,24 +14,20 @@ def prepare_comments(comments,result=Array.new, init=true, counter=0)
 		
 		result << "<div class=\"comments_left\"><p><span class=\"date\">#{time_ago_in_words(comment.created_at).upcase} AGO</span><br/> <span class=\"email\">#{comment.name}</span></p>" 
 		
-		if administrator? 
-			result << "<br/><span class=\"email\">#{comment.email}</span>" 
-		end
-		
+		result << "<br/><span class=\"email\">#{comment.email}</span>" if administrator? 
+			
 		result << "</div><div class=\"comments_right\"><h6>#{comment.subject}</h6>" 
 		
-		if administrator?  
-			result << "<p>" + link_to(comment.content.title, content_path(:id => comment.content.permalink)) + "</p>"
-		end
+		result << "<p>" + link_to(comment.content.title, content_path(:id => comment.content.permalink)) + "</p>" if administrator? 
 		
 		result << "<p>#{comment.body}</p></div>" 
 		
-		if administrator?
-			result << "<p>" + link_to_remote("Delete", :url => comment_path(:content_id => comment.content.id, :id => comment.id), :method => :delete) + "</p>"
-		end
+		result << "<p>" + link_to_remote("Delete", :url => comment_path(:content_id => comment.content.id, :id => comment.id), :method => :delete) + "</p>" if administrator? 
 					
 		result << 
-		"<div class=\"reply_comment\" id=\"reply_#{comment.id}\">" + link_to_remote("Reply to this comment", { :url => new_content_comment_child_comment_path(:content_id => comment.content.id, :comment_id => comment.id), :method => :get }, :href => "http://www.yahoo.com")  + "</div><div id=\"reply_form_#{comment.id}\"></div>"
+		"<div class=\"reply_comment\" id=\"reply_#{comment.id}\">" + 
+		link_to_remote("Reply to this comment", { :url => new_content_comment_child_comment_path(:content_id => comment.content.id, :comment_id => comment.id), :method => :get }, :href => "http://www.yahoo.com")  + 
+		"</div><div id=\"reply_form_#{comment.id}\"></div>"
 			
 		prepare_comments(comment.children,result,false,counter) unless comment.children.empty?
